ima: add ima-evm-rootfs.bbclass
Issue: TA226645

This class is used to sign the rootfs, which makes sure the produced
image(tar.bz2, WIC, etc) are signed by IMA key properly.

For upstream, it expects rootfs is installed by an installer, and during
build time of the installer, rootfs is signed and achived, and at runtime
of the installer, the signed rootfs is installed on the target.

So this class is for Wind River Linux only as we don't use that
installer metioned by meta-secure-core upstream. Instead, we sign the
rootfs at build time.

This class refers to
https://github.com/intel/meta-intel-iot-security/blob/master/meta-integrity/classes/ima-evm-rootfs.bbclass

(LOCAL REV: NOT UPSTREAM) -- WRL Specific

Signed-off-by: Yunguo Wei <yunguo.wei@windriver.com>